consolidate colorStringToUint32 code support hexcode colors in gradients
